+++ article-api/src/main/kotlin/com/sns/article/component/reaction/application/ReactionQueryService.kt
@@ -1,8 +1,8 @@
-package com.sns.article.reaction.application
+package com.sns.article.component.reaction.application
 
-import com.sns.article.reaction.domain.ReactionNotFoundException
-import com.sns.article.reaction.domain.ReactionRepository
-import com.sns.article.reaction.dto.ReactionDto
+import com.sns.article.component.reaction.domains.ReactionNotFoundException
+import com.sns.article.component.reaction.domains.ReactionRepository
+import com.sns.article.component.reaction.dtos.ReactionDto
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
