+++ article-api/src/main/kotlin/com/sns/article/component/reaction/domains/ReactionCreatedEvent.kt
@@ -1,3 +1,3 @@
-package com.sns.article.reaction.domain
+package com.sns.article.component.reaction.domains
 
 data class ReactionCreatedEvent(val reactionId: Long)
