+++ article-api/src/main/kotlin/com/sns/article/component/reaction/domains/ReactionNotFoundException.kt
@@ -1,3 +1,3 @@
-package com.sns.article.reaction.domain
+package com.sns.article.component.reaction.domains
 
 class ReactionNotFoundException : RuntimeException()
