+++ article-api/src/main/kotlin/com/sns/article/component/reaction/domains/ReactionRepository.kt
@@ -1,4 +1,4 @@
-package com.sns.article.reaction.domain
+package com.sns.article.component.reaction.domains
 
 import java.util.*
