+++ article-api/src/main/kotlin/com/sns/article/component/reaction/domains/ReactionTarget.kt
@@ -1,4 +1,4 @@
-package com.sns.article.reaction.domain
+package com.sns.article.component.reaction.domains
 
 import org.springframework.data.relational.core.mapping.Column
