+++ article-api/src/main/kotlin/com/sns/article/component/reaction/domains/ReactionTargetType.kt
@@ -1,4 +1,4 @@
-package com.sns.article.reaction.domain
+package com.sns.article.component.reaction.domains
 
 enum class ReactionTargetType {
     ARTICLE, COMMENT
