+++ article-api/src/main/kotlin/com/sns/article/component/reaction/domains/ReactionType.kt
@@ -1,0 +1,5 @@
+package com.sns.article.component.reaction.domains
+
+enum class ReactionType {
+    LIKE
+}
