+++ article-api/src/main/kotlin/com/sns/article/component/reaction/repositories/SpringDataJdbcReactionRepository.kt
@@ -1,7 +1,7 @@
-package com.sns.article.reaction.infra.repository
+package com.sns.article.component.reaction.repositories
 
-import com.sns.article.reaction.domain.Reaction
-import com.sns.article.reaction.domain.ReactionRepository
+import com.sns.article.component.reaction.domains.Reaction
+import com.sns.article.component.reaction.domains.ReactionRepository
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
