--- article-api/src/main/kotlin/com/sns/article/reaction/domain/ReactionType.kt
@@ -1,5 +1,0 @@
-package com.sns.article.reaction.domain
-
-enum class ReactionType {
-    LIKE
-}
